[𝘀𝗽𝗿] initial version
Created using spr 1.3.7
